RTBE-558 addressing code review comments

+++ src/main/java/com/rottentomatoes/movieapi/domain/model/TvAudienceReview.java
@@ -25,7 +25,7 @@ public class TvAudienceReview extends AbstractModel {
     public Boolean elite;
     public String mediaType;
     public Integer mediaId;
-    public Integer score;
+    public Double score;
     public String comment;
     public Integer commentLength;
     @JsonDeserialize(using = DeSerializeZonedDateTime.class)
